requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.48.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
pip-install==1.3.5
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
matplotlib==3.8.2
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
tifffile==2024.1.30
timm==0.9.12
torch
torch-dct
torchaudio
torchvision
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
yacs==0.1.8
zipp==3.17.0
fastdtw>=0.3.4
scipy>=1.12.0
